feat: split test into several test files
