+++ tests/TestDataService/TestDataService.spec.ts
@@ -1,10 +1,52 @@
-import { test, expect, type Product, Category, PropertyGroup, Customer, Manufacturer, PaymentMethod, Rule, ShippingMethod, Currency, Country } from '../../src/index';
+import {
+    test,
+    expect,
+    Customer,
+    PaymentMethod,
+    Rule,
+    ShippingMethod,
+    Currency,
+    Country,
+    PropertyGroup,
+    Product,
+    Manufacturer,
+    Category,
+    APIResponse,
+} from '../../src';
 
 test('Data Service', async ({
     TestDataService,
     AdminApiContext,
 }) => {
-    // Test test-data generation
+
+    const customer = await TestDataService.createCustomer({ firstName: 'Luke', lastName: 'Skywalker' })
+    expect(customer.firstName).toEqual('Luke');
+    expect(customer.lastName).toEqual('Skywalker');
+
+    const paymentMethod = await TestDataService.createBasicPaymentMethod({ name: 'Custom payment method' });
+    expect(paymentMethod.name).toEqual('Custom payment method');
+
+    const paymentMethodWithImage = await TestDataService.createPaymentMethodWithImage();
+    expect(paymentMethodWithImage.media).toBeDefined();
+
+    const rule = await TestDataService.createBasicRule({ description: 'Rule description' });
+    expect(rule.description).toEqual('Rule description');
+
+    const basicShippingMethod = await TestDataService.createBasicShippingMethod({ name: 'Custom shipping method' });
+    expect(basicShippingMethod.name).toEqual('Custom shipping method');
+
+    const shippingMethodWithImage = await TestDataService.createShippingMethodWithImage();
+    expect(shippingMethodWithImage.media).toBeDefined();
+
+    const currency = await TestDataService.createCurrency({ taxFreeFrom: 10 });
+    expect(currency.taxFreeFrom).toEqual(10);
+
+    const country = await TestDataService.createCountry();
+    expect(country.name).toBeDefined();
+
+    const customerGroup = await TestDataService.createCustomerGroup({ name: 'Custom customer group'});
+    expect(customerGroup.name).toEqual('Custom customer group');
+
     const category = await TestDataService.createCategory({ name: 'Custom Category' });
     expect(category.name).toEqual('Custom Category');
 
@@ -18,58 +60,6 @@ test('Data Service', async ({
     const propertyGroup = await TestDataService.createColorPropertyGroup();
     expect(propertyGroup.description).toEqual('Color');
 
-    const customer = await TestDataService.createCustomer({ firstName: 'Luke', lastName: 'Skywalker' })
-    expect(customer.firstName).toEqual('Luke');
-    expect(customer.lastName).toEqual('Skywalker');
-
-    const customerAddress = await TestDataService.getCustomerAddress(customer.defaultBillingAddressId);
-    const deliveryStateMachine = await TestDataService.getDeliveryStateMachine();
-    const deliveryState = await TestDataService.getStateMachineState(deliveryStateMachine.id);
-    const shippingMethod = await TestDataService.getShippingMethod();
-    const deliveryStruct = TestDataService.getBasicOrderDeliveryStruct(deliveryState, shippingMethod, customerAddress);
-
-    const shippingCosts = 12.99;
-    const totalPrice = 50 + shippingCosts;
-
-    // eslint-disable-next-line playwright/no-conditional-in-test
-    if (deliveryStruct.shippingCosts != null) {
-        deliveryStruct.shippingCosts.unitPrice = shippingCosts;
-        deliveryStruct.shippingCosts.totalPrice = shippingCosts;
-    }
-
-    const customShippingCosts = {
-        price: {
-            totalPrice: totalPrice,
-            positionPrice: totalPrice,
-            rawTotal: totalPrice,
-            netPrice: totalPrice,
-            taxStatus: 'gross',
-            calculatedTaxes: [{
-                tax: 0,
-                taxRate: 0,
-                price: totalPrice,
-            }],
-            taxRules: [{
-                taxRate: 0,
-                percentage: 100,
-            }],
-        },
-        deliveries: [deliveryStruct],
-    };
-
-    const orderWithCustomShippingCosts = await TestDataService.createOrder(
-        [{ product, quantity: 5 }],
-        customer,
-        customShippingCosts,
-    );
-    expect(orderWithCustomShippingCosts.price.totalPrice).toEqual(62.99);
-
-    const paymentMethod = await TestDataService.createBasicPaymentMethod({ name: 'Custom payment method' });
-    expect(paymentMethod.name).toEqual('Custom payment method');
-
-    const paymentMethodWithImage = await TestDataService.createPaymentMethodWithImage();
-    expect(paymentMethodWithImage.media).toBeDefined();
-
     const manufacturer = await TestDataService.createBasicManufacturer({ description: 'Test Description Manufacturer' });
     expect(manufacturer.description).toEqual('Test Description Manufacturer');
 
@@ -79,69 +69,30 @@ test('Data Service', async ({
     await TestDataService.assignManufacturerProduct(manufacturer.id, product.id)
     expect(product.manufacturerId).toBeDefined();
 
-    const rule = await TestDataService.createBasicRule({ description: 'Rule description' });
-    expect(rule.description).toEqual('Rule description');
-
-    const parentProduct = await TestDataService.createBasicProduct();
-    const propertyGroups: PropertyGroup[] = [];
-    const propertyGroupColor = await TestDataService.createColorPropertyGroup();
-    const propertyGroupText = await TestDataService.createTextPropertyGroup();
-    propertyGroups.push(propertyGroupColor);
-    propertyGroups.push(propertyGroupText);
-
-    const variantProducts = await TestDataService.createVariantProducts(parentProduct, propertyGroups, { description: 'Variant description' });
-    expect(variantProducts.length).toEqual(9);
-    expect(variantProducts[0].description).toEqual('Variant description');
-
-    const basicShippingMethod = await TestDataService.createBasicShippingMethod({ name: 'Custom shipping method' });
-    expect(basicShippingMethod.name).toEqual('Custom shipping method');
-
-    const shippingMethodWithImage = await TestDataService.createShippingMethodWithImage();
-    expect(shippingMethodWithImage.media).toBeDefined();
-
     const cmsType = 'product_detail';
     const cmsPageName = 'Custom product detail page';
     const cmsPage = await TestDataService.createBasicPageLayout(cmsType, { name: cmsPageName });
     expect(cmsPage.name).toEqual(cmsPageName);
     expect(cmsPage.type).toEqual(cmsType);
 
-    const currency = await TestDataService.createCurrency({ taxFreeFrom: 10 });
-    expect(currency.taxFreeFrom).toEqual(10);
-
-    const country = await TestDataService.createCountry();
-    expect(country.name).toBeDefined();
-
-    const customerGroup = await TestDataService.createCustomerGroup({ name: 'Custom customer group'});
-    expect(customerGroup.name).toEqual('Custom customer group');
+    const parentProduct = await TestDataService.createBasicProduct();
+    const propertyGroupColor = await TestDataService.createColorPropertyGroup();
+    const propertyGroupText = await TestDataService.createTextPropertyGroup();
+    const propertyGroups: PropertyGroup[] = [];
+    propertyGroups.push(propertyGroupColor);
+    propertyGroups.push(propertyGroupText);
+    const variantProducts = await TestDataService.createVariantProducts(parentProduct, propertyGroups, { description: 'Variant description' });
+    expect(variantProducts.length).toEqual(9);
+    expect(variantProducts[0].description).toEqual('Variant description');
 
     // Test data clean-up with deactivated cleansing process
     TestDataService.setCleanUp(false);
     const cleanUpFalseResponse = await TestDataService.cleanUp();
     expect(cleanUpFalseResponse).toBeNull();
 
-    const categoryResponse = await AdminApiContext.get(`./category/${category.id}?_response=detail`);
-    const { data: databaseCategory } = (await categoryResponse.json()) as { data: Category };
-    expect(databaseCategory.id).toBe(category.id);
-
-    const productResponse = await AdminApiContext.get(`./product/${product.id}?_response=detail`);
-    const { data: databaseProduct } = (await productResponse.json()) as { data: Product };
-    expect(databaseProduct.id).toBe(product.id);
-
-    const digitalProductResponse = await AdminApiContext.get(`./product/${digitalProduct.id}?_response=detail`);
-    const { data: databaseDigitalProduct } = (await digitalProductResponse.json()) as { data: Product };
-    expect(databaseDigitalProduct.id).toBe(digitalProduct.id);
-
-    const propertyGroupResponse = await AdminApiContext.get(`./property-group/${propertyGroup.id}?_response=detail`);
-    const { data: databasePropertyGroup } = (await propertyGroupResponse.json()) as { data: PropertyGroup };
-    expect(databasePropertyGroup.id).toBe(propertyGroup.id);
-
     const customerResponse = await AdminApiContext.get(`./customer/${customer.id}?_response=detail`);
     const { data: databaseCustomer } = (await customerResponse.json()) as { data: Customer };
     expect(databaseCustomer.id).toBe(customer.id);
-
-    const manufacturerResponse = await AdminApiContext.get(`./product-manufacturer/${manufacturer.id}?_response=detail`);
-    const { data: databaseManufacturer } = (await manufacturerResponse.json()) as { data: Manufacturer };
-    expect(databaseManufacturer.id).toBe(manufacturer.id);
 
     const paymentMethodResponse = await AdminApiContext.get(`./payment-method/${paymentMethod.id}?_response=detail`);
     const { data: databasePaymentMethod } = (await paymentMethodResponse.json()) as { data: PaymentMethod };
@@ -163,29 +114,44 @@ test('Data Service', async ({
     const { data: databaseCountry } = (await countryResponse.json()) as { data: Country };
     expect(databaseCountry.id).toBe(country.id);
 
+    const productResponse = await AdminApiContext.get(`./product/${product.id}?_response=detail`);
+    const { data: databaseProduct } = (await productResponse.json()) as { data: Product };
+    expect(databaseProduct.id).toBe(product.id);
+
+    const digitalProductResponse = await AdminApiContext.get(`./product/${digitalProduct.id}?_response=detail`);
+    const { data: databaseDigitalProduct } = (await digitalProductResponse.json()) as { data: Product };
+    expect(databaseDigitalProduct.id).toBe(digitalProduct.id);
+
+    const propertyGroupResponse = await AdminApiContext.get(`./property-group/${propertyGroup.id}?_response=detail`);
+    const { data: databasePropertyGroup } = (await propertyGroupResponse.json()) as { data: PropertyGroup };
+    expect(databasePropertyGroup.id).toBe(propertyGroup.id);
+
+    const manufacturerResponse = await AdminApiContext.get(`./product-manufacturer/${manufacturer.id}?_response=detail`);
+    const { data: databaseManufacturer } = (await manufacturerResponse.json()) as { data: Manufacturer };
+    expect(databaseManufacturer.id).toBe(manufacturer.id);
+
+    const categoryResponse = await AdminApiContext.get(`./category/${category.id}?_response=detail`);
+    const { data: databaseCategory } = (await categoryResponse.json()) as { data: Category };
+    expect(databaseCategory.id).toBe(category.id);
+
     // Test data clean-up with activated cleansing process
     TestDataService.setCleanUp(true);
-    const cleanUpResponse = await TestDataService.cleanUp();
-
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
+    const cleanUpResponse = await TestDataService.cleanUp() as APIResponse;
     expect(cleanUpResponse.ok()).toBeTruthy();
 
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
     const cleanUp = await cleanUpResponse.json();
     expect(cleanUp['notFound'].length).toBe(0);
-    expect(cleanUp['deleted']['category']).toBeDefined();
     expect(cleanUp['deleted']['media']).toBeDefined();
-    expect(cleanUp['deleted']['property_group']).toBeDefined();
-    expect(cleanUp['deleted']['property_group_option']).toBeDefined();
     expect(cleanUp['deleted']['customer']).toBeDefined();
-    expect(cleanUp['deleted']['product_manufacturer']).toBeDefined();
     expect(cleanUp['deleted']['payment_method']).toBeDefined();
     expect(cleanUp['deleted']['shipping_method']).toBeDefined();
     expect(cleanUp['deleted']['rule']).toBeDefined();
-    expect(cleanUp['deleted']['cms_page']).toBeDefined();
     expect(cleanUp['deleted']['currency']).toBeDefined();
     expect(cleanUp['deleted']['country']).toBeDefined();
     expect(cleanUp['deleted']['customer_group']).toBeDefined();
+    expect(cleanUp['deleted']['category']).toBeDefined();
+    expect(cleanUp['deleted']['property_group']).toBeDefined();
+    expect(cleanUp['deleted']['property_group_option']).toBeDefined();
+    expect(cleanUp['deleted']['product_manufacturer']).toBeDefined();
+    expect(cleanUp['deleted']['cms_page']).toBeDefined();
 });
